refactor(internal): move gensnippets into gapicgen
