+++ internal/gapicgen/gensnippets/gensnippets.go
