+++ internal/gapicgen/go.mod
@@ -3,7 +3,8 @@ module cloud.google.com/go/internal/gapicgen
 go 1.16
 
 require (
-	cloud.google.com/go/internal/gensnippets v0.0.0-00010101000000-000000000000
+	cloud.google.com/go v0.81.0
+	cloud.google.com/go/internal/godocfx v0.0.0-20210416164805-eb2ed0dc7d6b
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/go-github/v34 v34.0.0
 	github.com/kr/text v0.2.0 // indirect
@@ -12,11 +13,8 @@ require (
 	github.com/shurcooL/graphql v0.0.0-20181231061246-d48a9a75455f // indirect
 	golang.org/x/oauth2 v0.0.0-20210413134643-5e61552d6c78
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	golang.org/x/sys v0.0.0-20210415045647-66c3f260301c
 	gopkg.in/check.v1 v1.0.0-20200227125254-8fa46927fb4f // indirect
 	gopkg.in/src-d/go-git.v4 v4.13.1
 	gopkg.in/yaml.v2 v2.4.0
 )
-
-replace cloud.google.com/go/internal/gensnippets => ../gensnippets
-
-replace cloud.google.com/go/internal/godocfx => ../godocfx
