--- internal/gensnippets/go.mod
@@ -1,13 +1,0 @@
-module cloud.google.com/go/internal/gensnippets
-
-go 1.16
-
-replace cloud.google.com/go/internal/godocfx => ../godocfx
-
-replace cloud.google.com/go => ../..
-
-require (
-	cloud.google.com/go v0.81.0
-	cloud.google.com/go/internal/godocfx v0.0.0-00010101000000-000000000000
-	golang.org/x/sys v0.0.0-20210403161142-5e06dd20ab57
-)
